requirements.txt
astroid==2.12.5
async-generator==1.10
attrs==22.1.0
boto3==1.24.62
botocore==1.27.62
certifi==2022.6.15
charset-normalizer==2.1.1
dill==0.3.5.1
dnspython==2.2.1
h11==0.13.0
idna==3.3
isort==5.10.1
jmespath==1.0.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
outcome==1.2.0
pathlib==1.0.1
Pillow==9.2.0
platformdirs==2.5.2
pymongo==4.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.28.1
s3transfer==0.6.0
selenium==4.4.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tomli==2.0.1
tomlkit==0.11.4
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
webdriver-manager==3.8.3
wrapt==1.14.1
wsproto==1.2.0
